+++ CHANGELOG.md
@@ -23,6 +23,8 @@ In this release we had a little focus on:
   and expanentional numbers
 - Enforces to use `1e10` instead of `1e+10`
 - Enforces to use big letters for hex numbers: `0xAB` instead of `0xab`
+- Enforces to use `r'\n'` instead of `'\\n'`
+- Forbids to have unicode escape characters inside binary strings
 
 ### Bugfixes